move theme and part traits to their own files
